+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 - Add dark mode of Banner component [#101](https://github.com/mishka-group/mishka_chelekom/pull/101)
 - Add dark mode of Breadcrumb component [#102](https://github.com/mishka-group/mishka_chelekom/pull/102)
 - Add dark mode of Blockquote component [#103](https://github.com/mishka-group/mishka_chelekom/pull/103)
+- Add dark mode of dropdown component [#104](https://github.com/mishka-group/mishka_chelekom/pull/104)
 
 ### Bugs:
 - Fix un-CSP progress mounted [#72](https://github.com/mishka-group/mishka_chelekom/pull/72)